eth-accounts module simplified and SendTransactionMethod handlin in core-method module updated. Anything is still in a POC phase
